Prepare for minor update with some extra features
Sign restrictions are now computed iteratively, which should
improve the speed at which they are found. The limit of draws
is still rather modest, to avoid long computation times by default.
New features are hist_decomp, rmse, lps, and WAIC.
Lastly, the hope is to fix the vignette's tangle output for CRAN.